Use UTC datetimes internally
Python datetime object comparisons and math don't work as expected when
the objects are aware and have the same tzinfo attribute. Basically the
fold attribute is ignored in this case, which can lead to wrong results.
Avoid this problem by using aware times in UTC internally. Only use the
location's tz for user visible attributes.